tools/github_repo_management/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = 'ff_repo'
requires-python = '>=3.6'
readme = 'README.md'
dependencies = [
  "pygithub",
  "frozendict",
  "sqlitedict"
]

[tool.black]
line-length = 120
target-version = ['py36']

[tool.ruff]
line-length = 120
target-version = "py37"

[tool.mypy]
warn_unused_configs = true
enable_recursive_aliases = true
python_version = '3.6'

[[tool.mypy.overrides]]
module = "sqlitedict.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "setuptools.*"
ignore_missing_imports = true
